automation workflow (#114)
* feat: Add GitHub workflow for automating native agent updates

This introduces a new GitHub Actions workflow to automate the process of updating the native Android and iOS agent versions.

* feat: add release automation workflow

Adds a GitHub Actions workflow to automate the creation of releases and the updating of documentation.

This workflow is triggered automatically when a pull request with "Update native agents" in the title is merged into the `master` branch, or it can be run manually.

The workflow performs the following actions:
- Creates a new Git tag and a corresponding GitHub Release.
- Populates the release notes from `CHANGELOG.md`.
- Automatically generates a pull request to the `docs-website` repository with the new release notes.

Additionally, new documentation for the workflows (`README.md` and `WORKFLOW_SETUP.md`) has been added to explain the setup and usage of the automation.